Move BaseQuery into query module and add session directly to the query.
